+++ source/ftrack_api/entity/location.py
@@ -17,6 +17,13 @@ from ftrack_api.logging import LazyLogMessage as L
 from future.utils import with_metaclass
 
 
+MixinBaseClass = with_metaclass(
+      ftrack_api.entity.base.DynamicEntityTypeMetaclass,
+      ftrack_api.entity.base._EntityBase,
+      collections.MutableMapping
+)
+
+
 class Location(ftrack_api.entity.base.Entity):
     '''Represent storage for components.'''
 
@@ -613,12 +620,7 @@ class Location(ftrack_api.entity.base.Entity):
         return self.accessor.get_url(resource_identifier)
 
 
-class MemoryLocationMixin(
-    with_metaclass(
-        ftrack_api.entity.base.DynamicEntityTypeMetaclass, 
-        ftrack_api.entity.base._EntityBase, 
-        collections.MutableMapping
-    )):
+class MemoryLocationMixin(MixinBaseClass):
     '''Represent storage for components.
 
     Unlike a standard location, only store metadata for components in this
@@ -686,12 +688,7 @@ class MemoryLocationMixin(
         return resource_identifiers
 
 
-class UnmanagedLocationMixin(    
-    with_metaclass(
-        ftrack_api.entity.base.DynamicEntityTypeMetaclass, 
-        ftrack_api.entity.base._EntityBase, 
-        collections.MutableMapping
-    )):
+class UnmanagedLocationMixin(MixinBaseClass):
     '''Location that does not manage data.'''
 
     def _add_data(self, component, resource_identifier, source):
@@ -728,12 +725,7 @@ class OriginLocationMixin(MemoryLocationMixin, UnmanagedLocationMixin):
         return context
 
 
-class ServerLocationMixin(    
-    with_metaclass(
-        ftrack_api.entity.base.DynamicEntityTypeMetaclass, 
-        ftrack_api.entity.base._EntityBase, 
-        collections.MutableMapping
-    )):
+class ServerLocationMixin(MixinBaseClass):
     '''Location representing ftrack server.
 
     Adds convenience methods to location, specific to ftrack server.