+++ test/unit/event/test_hub.py
@@ -9,6 +9,7 @@ import subprocess
 import sys
 
 import pytest
+from flaky import flaky
 
 import ftrack_api.event.hub
 import ftrack_api.event.subscriber
@@ -442,6 +443,7 @@ def test_unsubscribe(event_hub):
     assert_callbacks(event_hub, [callback_b])
 
 
+@flaky(max_runs=2, min_passes=1)
 def test_unsubscribe_whilst_disconnected(event_hub):
     '''Unsubscribe whilst disconnected.'''
     identifier = event_hub.subscribe('topic=test', None)