+++ source/ftrack_api/event/hub.py
@@ -46,19 +46,8 @@ ServerDetails = collections.namedtuple('ServerDetails', [
 ])
 
 
-
-
 class EventHub(object):
     '''Manage routing of events.'''
-
-    _future_signature_warning = (
-        'When constructing your Session object you did not explicitly define '
-        'auto_connect_event_hub as True even though you appear to be publishing '
-        'and / or subscribing to asynchronous events. In version version 2.0 of '
-        'the ftrack-python-api the default behavior will change from True '
-        'to False. Please make sure to update your tools. You can read more at '
-        'http://ftrack-python-api.rtd.ftrack.com/en/stable/release/migration.html'
-    )
 
     def __init__(self, server_url, api_user, api_key):
         '''Initialise hub, connecting to ftrack *server_url*.
@@ -92,8 +81,6 @@ class EventHub(object):
         # disconnection. Equates to 5 minutes.
         self._auto_reconnect_attempts = 30
         self._auto_reconnect_delay = 10
-
-        self._deprecation_warning_auto_connect = False
 
         # Mapping of Socket.IO codes to meaning.
         self._code_name_mapping = {
@@ -153,8 +140,6 @@ class EventHub(object):
         connected or connection fails.
 
         '''
-
-        self._deprecation_warning_auto_connect = False
 
         if self.connected:
             raise ftrack_api.exception.EventHubConnectionError(
@@ -574,10 +559,6 @@ class EventHub(object):
         event will be caught by this method and ignored.
 
         '''
-        if self._deprecation_warning_auto_connect and not synchronous:
-            warnings.warn(
-                self._future_signature_warning, FutureWarning
-            )
 
         try:
             return self._publish(
@@ -737,10 +718,6 @@ class EventHub(object):
             # Automatically publish a non None response as a reply when not in
             # synchronous mode.
             if not synchronous:
-                if self._deprecation_warning_auto_connect:
-                    warnings.warn(
-                        self._future_signature_warning, FutureWarning
-                    )
 
                 if response is not None:
                     try: