+++ doc/release/release_notes.rst
@@ -15,6 +15,16 @@ Release Notes
     
         Provide support for python3.X.
 
+    .. change:: changed
+        :tags: Session
+
+        Private method :meth:`Session._entity_reference` has been deprecated.
+
+        .. note::
+
+            You should now use the public method :meth:`Session.entity_reference`.
+
+
 .. release:: 1.8.1
     :date: 2019-10-30