+++ doc/release/release_notes.rst
@@ -24,6 +24,15 @@ Release Notes
 
             You should now use the public method :meth:`Session.entity_reference`.
 
+    .. change:: changed
+        :tags: Session
+        
+        Private method :meth:`Session._call` has been deprecated.
+
+        .. note::
+
+            You should now use the public method :meth:`Session.call`.
+
 
 .. release:: 1.8.1
     :date: 2019-10-30