+++ doc/release/release_notes.rst
@@ -16,13 +16,31 @@ Release Notes
         Provide support for python3.X.
 
     .. change:: changed
-        :tag: session
+        :tags: Session
+
+        Private method :meth:`Session._entity_reference` has been deprecated.
+
+        .. note::
+
+            You should now use the public method :meth:`Session.entity_reference`.
+
+    .. change:: changed
+        :tags: Session
+        
+        Private method :meth:`Session._call` has been deprecated.
+
+        .. note::
+
+            You should now use the public method :meth:`Session.call`.
+
+    .. change:: changed
+        :tag: Session
 
         Public method :meth:`Session.delayed_job` has been deprecated.
 
         .. note:: 
 
-            Use :meth:`Session.call` instead.
+           You should now use :meth:`Session.call`.
  
 .. release:: 1.8.1
     :date: 2019-10-30