+++ doc/release/release_notes.rst
@@ -47,11 +47,15 @@ Release Notes
 
         LocationMixins are not compatible with Python 3.
     
+    .. change:: fixed
+        :tag: Test
+
+        Entity test fails due to missing parents.
+
     .. change:: new
         :tag: Tests
 
         Add support for flaky tests to improve test reliability.
-
 
 
 .. release:: 1.8.1